requirements.txt
flask
redis
xmltodict
beautifulsoup4
requests
redlock-py
